Fix the bugs found in static analysis
This commit fixes the static analyzer reported issues:

Operands don't affect result
some conditions are not required to check as its always true

+++ src/bridgingcommands.cpp
@@ -106,7 +106,7 @@ static inline bool ipmbDataChecksumValidate(const ipmbHeader* ipmbHeader,
                                 (length - ipmbConnectionHeaderLength));
 }
 
-static bool isFrameValid(const ipmbHeader* frame, uint8_t length)
+static bool isFrameValid(const ipmbHeader* frame, size_t length)
 {
     if ((length < ipmbMinFrameLength) || (length > ipmbMaxFrameLength))
     {
